Fix factory not resolving in any direction other than south/north
The setBlockState was being called by onBlockPlacedBy after the te was updated. setBlockState was destroying the TE and creating a new one, therefore losing the facing setting. Moved the setBlockState before the TE update.

Need to look at storing facing in either the TE or the block state, not both!