Use 3 fakeplayers to handle the different looting levels [#47]
When a fakeplayer equips a weapon an equip sound is played. We only
have one fakeplayer which is used on all factories, so the weapon can
be changed frequently. To stop this from happening we create 3 fakeplayers
each holding a different level of looting sword and never change their
equipment once loaded. The user may hear one equip sound for each
initial fakeplayer creation.