Replace unknown loot drops with dirt.
If there is a mapping in the loaded loot file that is not valid,
eg. the mod has been uninstalled, then keep processing but replace
that item with dirt.

This should really be replaced with a Woot "bad item", so it will
never generate in the output, unlike dirt. ie. I would blacklist
the Woot "bad item".