When learning pickup mob drops not using LivingDropsEvent [#75]
Some mods use the LivingDropsEvent, but don't add the drop to the
event drop list. They directly spawn the item in the world. This
change looks for items in the immediate area of the factory when
learning, to try and pickup and learn these items. This means eg.
that ArmorPlus mod drops will be hoovered up and learned.