Handle when loot.json exists but is empty [#153]
If loot.json exist then we load it on startup. If there is any
error in it then GSON.fromJson throws the JsonParseException
which we catch and handle.
However loot.json exists and is EMPTY, then nothing is thrown
and the lootmap return is null. This causes crashes later
on when lootmap is accessed as lootmap is now null.
Therefore we check for lootmap == null after the load of the file
to handle the exists but is empty case.