[#257] Don't learn mobs that only use custom drops
Some mobs ie. Ender Dragon, don't die in a standard way. Therefore
we have custom drops in the config file for those mobs. Need to ensure
that the learning process is not applied to those mobs.
Currently this only applies to the Ender Dragon.

This is different from using the custom drops to apply extra drops
to a normal mob.