Fix crash when opening gui on unformed factory [#297]
When a factory is activate the gui will try to open. However
on opening it will then request the factory config from the server.
If the factory is not formed then there will be no tier set and
the packet creating will exception.

Check that the factory is formed before opening the gui