Change from String.format to fixed string
The use of the String.format is unnecessary as I'm only adding
one message. The reason it was an issue is that '%f' expands to use
different seperator characters depending on locale. So UK/US may
use  '.' but Germany uses ',' which breaks the IMC decode.

There is the option to specify the locale to make the seperator
consistent but I'll just swap over to a fix string.