infuser: internal tanks can toggle between dump and strict
Strict tanks will not process the input if the complete dye output
cannot be stored in the internal tanks. Dump mode will keep processing
and dump the excess. This can be used when one or more internal
tanks are full, but the input item could be filling those tanks
and one of the other tanks with space. eg. purple in, but the blue
tank is full, will still allow the input to be processed and the
red tank to fill.