Fix heart cell capacity showing wrong value
This is down to me doing the int tracking for a non-short
incorrectly. You must have get/set operating on the save
value and for the cell I get from one and set from another.
This works for shorts but clearly not for ints where
the value is passed over in type values. So shorts
are just working out of luck!